requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.5.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==21.2.0  # Necesario para producción
whitenoise==6.6.0  # Para servir archivos estáticos
django-environ==0.11.2  # Mejor manejo de variables de entorno
